deleted:    .env.example 	new file:   Portofolio 	new file:   app/Exports/BeasiswaExport.php 	new file:   app/Exports/KegiatanExport.php 	new file:   app/Exports/LaporanExport.php 	new file:   app/Exports/PrestasiExport.php 	new file:   app/Http/Controllers/BeritaPublicController.php 	modified:   app/Http/Controllers/DashboardController.php 	new file:   app/Http/Controllers/LaporanController.php 	modified:   app/Http/Controllers/LoginController.php 	new file:   app/Http/Controllers/User/UserBeasiswaController.php 	new file:   app/Http/Controllers/User/UserKegiatanController.php 	new file:   app/Http/Controllers/User/UserPrestasiController.php 	new file:   app/Http/Controllers/User/UserProfileController.php 	new file:   app/Http/Controllers/admin/BeasiswaController.php 	new file:   app/Http/Controllers/admin/BeritaController.php 	modified:   app/Http/Controllers/admin/DashboardController.php 	new file:   app/Http/Controllers/admin/KegiatanController.php 	modified:   app/Http/Controllers/admin/LoginController.php 	new file:   app/Http/Controllers/admin/PrestasiController.php 	new file:   app/Http/Controllers/admin/profileController.php 	modified:   app/Http/Middleware/AdminAuthenticate.php 	modified:   app/Http/Middleware/AdminRedirect.php 	new file:   app/Models/Admin.php 	new file:   app/Models/Beasiswa.php 	new file:   app/Models/Berita.php 	new file:   app/Models/Kegiatan.php 	new file:   app/Models/Prestasi.php 	modified:   app/Models/User.php 	new file:   app/Models/mahasiswa.php 	new file:   app/Models/user/UserPrestasi.php 	modified:   composer.json 	modified:   composer.lock 	modified:   config/app.php 	modified:   database/migrations/0001_01_01_000000_create_users_table.php 	modified:   database/migrations/2024_08_13_102039_add_users_table.php 	new file:   database/migrations/2024_10_28_150944_create_prestasis_table.php 	new file:   database/migrations/2024_10_28_150951_create_beasiswas_table.php 	new file:   database/migrations/2024_10_28_150956_create_kegiatans_table.php 	new file:   database/migrations/2024_11_18_153719_create_mahasiswas_table.php 	new file:   database/migrations/2024_12_01_133808_add_status_to_beasiswas_table.php 	new file:   database/migrations/2024_12_04_164301_drop_status_from_beasiswas_table.php 	new file:   database/migrations/2024_12_10_081918_create_beritas_table.php 	new file:   database/migrations/2024_12_14_094437_create_admins_table.php 	new file:   public/images/berita1.jpg 	new file:   public/images/berita2.jpg 	new file:   public/images/berita3.jpg 	new file:   public/images/bg.jpg 	new file:   public/images/bg22.png 	new file:   public/images/prestasi.png 	new file:   public/images/prestasi2.jpg 	new file:   public/images/prestasi3.jpg 	new file:   resources/css/style.css 	new file:   resources/views/account/beasiswa/create.blade.php 	new file:   resources/views/account/beasiswa/index.blade.php 	modified:   resources/views/account/dashboard.blade.php 	new file:   resources/views/account/forgot-password.blade.php 	new file:   resources/views/account/kegiatan/create.blade.php 	new file:   resources/views/account/kegiatan/index.blade.php 	new file:   resources/views/account/layouts/app.blade.php 	new file:   resources/views/account/layouts/footer.blade.php 	new file:   resources/views/account/layouts/navbar.blade.php 	modified:   resources/views/account/login.blade.php 	new file:   resources/views/account/prestasi/create.blade.php 	new file:   resources/views/account/prestasi/index.blade.php 	new file:   resources/views/account/profile.blade.php 	modified:   resources/views/account/register.blade.php 	new file:   resources/views/admin/beasiswa/create.blade.php 	new file:   resources/views/admin/beasiswa/edit.blade.php 	new file:   resources/views/admin/beasiswa/index.blade.php 	new file:   resources/views/admin/beasiswa/pdf.blade.php 	new file:   resources/views/admin/berita/create.blade.php 	new file:   resources/views/admin/berita/edit.blade.php 	new file:   resources/views/admin/berita/index.blade.php 	new file:   resources/views/admin/berita/show.blade.php 	modified:   resources/views/admin/dashboard.blade.php 	new file:   resources/views/admin/kegiatan/create.blade.php 	new file:   resources/views/admin/kegiatan/edit.blade.php 	new file:   resources/views/admin/kegiatan/index.blade.php 	new file:   resources/views/admin/kegiatan/pdf.blade.php 	new file:   resources/views/admin/laporan/index.blade.php 	new file:   resources/views/admin/laporan/pdf.blade.php 	new file:   resources/views/admin/layouts/crud.blade.php 	modified:   resources/views/admin/login.blade.php 	new file:   resources/views/admin/prestasi/create.blade.php 	new file:   resources/views/admin/prestasi/edit.blade.php 	new file:   resources/views/admin/prestasi/index.blade.php 	new file:   resources/views/admin/prestasi/pdf.blade.php 	new file:   resources/views/admin/profile.blade.php 	new file:   resources/views/admin/register.blade.php 	new file:   resources/views/admin/reset.blade.php 	modified:   resources/views/welcome.blade.php 	modified:   routes/web.php

+++ database/migrations/2024_08_13_102039_add_users_table.php
@@ -11,9 +11,12 @@ return new class extends Migration
      */
     public function up(): void
     {
-        Schema::table('users', function($table) {
-            $table->enum('role',['customer','admin'])->default('customer');
-        });
+        // Cek jika kolom 'role' belum ada
+        if (!Schema::hasColumn('users', 'role')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->enum('role', ['customer', 'admin'])->default('customer');
+            });
+        }
     }
 
     /**
@@ -21,8 +24,11 @@ return new class extends Migration
      */
     public function down(): void
     {
-        Schema::table('users', function($table) {
-            $table->dropColumn('role');
-        });
+        // Cek jika kolom 'role' ada dan hapus
+        if (Schema::hasColumn('users', 'role')) {
+            Schema::table('users', function (Blueprint $table) {
+                $table->dropColumn('role');
+            });
+        }
     }
 };
